Changed ChamberOfCommerce datatype to decimal, with 8 numbers and 0 decimals

+++ FSBeheer/FSBeheer/Model/Customer.cs
@@ -22,7 +22,7 @@ namespace FSBeheer.Model
         public string City { get; set; }
         public string ZipCode { get; set; }
         public DateTime? StartingDate { get; set; }
-        public short? ChamberOfCommerceNumber { get; set; }
+        public decimal? ChamberOfCommerceNumber { get; set; }
         public virtual ObservableCollection<Contact> Contacts { get; set; }
         public virtual ObservableCollection<Event> Events { get; set; }
         public bool IsDeleted { get; set; }
